Add files via upload
Approach
HTML: The design offers an easy-to-use form on which consumers can choose and enter the quantity of each kind of chocolate.
CSS: With an emphasis on usability and clarity, simple and responsive styling makes sure the form is accessible on a variety of devices.
JavaScript: As the user modifies their selections, this script dynamically computes the total items and price. It gives instant feedback on the current total cost and enforces the 8 item restriction.